+++ src/uml_socket/device.c
@@ -41,8 +41,8 @@ static char *device_info;
 extern char *identname;
 extern bool running;
 
-static int device_total_in = 0;
-static int device_total_out = 0;
+static uint64_t device_total_in = 0;
+static uint64_t device_total_out = 0;
 
 enum request_type { REQ_NEW_CONTROL };
 
@@ -272,6 +272,6 @@ bool write_packet(vpn_packet_t *packet) {
 
 void dump_device_stats(void) {
 	logger(LOG_DEBUG, "Statistics for %s %s:", device_info, device);
-	logger(LOG_DEBUG, " total bytes in:  %10d", device_total_in);
-	logger(LOG_DEBUG, " total bytes out: %10d", device_total_out);
+	logger(LOG_DEBUG, " total bytes in:  %10"PRIu64, device_total_in);
+	logger(LOG_DEBUG, " total bytes out: %10"PRIu64, device_total_out);
 }
